refactor(guess_sequence_solver): Return a decision tree
This might also be considered a fix since the previous implementation was not behaving in a way that was useful.

The signature wasn't practical because there is no sequence of constraints that the player is guaranteed to get correct 100% of the time.

Using a decision tree as the result is more practical since the player can follow the branch that corresponds with the reality of their guess.